Create Rubric
New web service, with one method. We will pass it a date time and it will count down the minutes to that date. In the client web site, add a couple of pictures and text. Then add an Ajax update panel, with a script manager and a  timer. Inside the update panel we will call the service once every minute and show the countdown.